+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * perf: Remove type guesser from `TypeResolver`.
 * perf(`ObjectCache`): Optimize using plain arrays.
 * perf: Optimize `ObjectCache` & `Context`.
+* perf(`TransformerRegistry`): Cache `TransformerInterface` instances.
 
 ## 0.5.14